Restructured music/tuning packages.
