+++ lib/src/main/java/com/rohankhayech/choona/lib/view/viewmodel/TunerViewModel.kt
@@ -21,9 +21,9 @@ package com.rohankhayech.choona.lib.view.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.rohankhayech.choona.lib.controller.tuner.Tuner
+import com.rohankhayech.choona.lib.controller.tunings.TuningList
+import com.rohankhayech.choona.lib.model.tuning.Tuning
 import com.rohankhayech.choona.lib.model.tuning.TuningEntry
-import com.rohankhayech.choona.lib.model.tuning.TuningList
-import com.rohankhayech.music.Tuning
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
