+++ lib/src/test/java/com/rohankhayech/choona/lib/model/tuning/NotesTest.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.rohankhayech.music;
+package com.rohankhayech.choona.lib.model.tuning;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
