+++ lib/src/test/java/com/rohankhayech/choona/lib/view/viewmodel/TunerViewModelTest.kt
@@ -18,8 +18,8 @@
 
 package com.rohankhayech.choona.lib.view.viewmodel
 
+import com.rohankhayech.choona.lib.model.tuning.Tuning
 import com.rohankhayech.choona.lib.model.tuning.TuningEntry
-import com.rohankhayech.music.Tuning
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.StandardTestDispatcher
